Flyttet komponenter ut av pages
