+++ apps/skde/src/components/HospitalProfile/ExpandableItemBox/index.tsx
