+++ apps/skde/src/components/HospitalProfile/HospitalProfileStyles/index.tsx
